billing-service/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.28.2
gunicorn==21.2.0
python-dotenv==1.0.0
reportlab==3.6.12  # For PDF generation
django-filter==23.1
django-cors-headers==4.0.0
dj-database-url==1.2.0